requirements.txt
httpx==0.25.2
fastapi==0.112.0
requests==2.31.0
urllib3==1.26.18
uvicorn==0.22.0
tz==0.2.2
python-dateutil==2.8.2
beautifulsoup4==4.12.2
tabulate==0.9.0
pyyaml>=6.0
torch>=1.9.0
numpy>=1.20.0
scikit-learn==1.6.1
matplotlib==3.10.1



